@@ -10,18 +10,16 @@
         "services/*"
     ],
     "devDependencies": {
-        "@dhis2/cli-app-scripts": "^10.2.0",
-        "@dhis2/cli-style": "^10.4.3",
+        "@dhis2/cli-app-scripts": "^12.0.0-alpha.21",
+        "@dhis2/cli-style": "10.4.3",
         "@dhis2/cli-utils-docsite": "^2.0.3",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^12.0.0",
-        "@testing-library/react-hooks": "^7.0.1",
+        "@testing-library/jest-dom": "^6.6.3",
+        "@testing-library/react": "^16.0.1",
         "@types/jest": "^24.9.0",
         "@types/lodash": "^4.14.171",
         "@types/node": "^13.1.8",
-        "@types/react": "^16.9.18",
-        "@types/react-dom": "^16.9.5",
-        "@types/testing-library__dom": "^7.5.0",
+        "@types/react": "^18.3.12",
+        "@types/react-dom": "^18.3.1",
         "@typescript-eslint/eslint-plugin": "^5.47.0",
         "@typescript-eslint/parser": "^5.47.0",
         "concurrently": "^5.0.2",
@@ -31,9 +29,8 @@
         "idb": "^6.1.3",
         "loop": "^3.3.4",
         "prop-types": "^15.7.2",
-        "react": "^16.12.0",
-        "react-dom": "^16.12.0",
-        "react-test-renderer": "^16.12",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
         "rimraf": "^3.0.0",
         "typescript": "^4.0.5"
     },
