@@ -1,4 +1,4 @@
-import { renderHook, act } from '@testing-library/react-hooks'
+import { renderHook, act } from '@testing-library/react'
 import React, { ReactNode } from 'react'
 import { AlertsProvider, useAlerts, useAlert } from '../index'
 
